replacing the moneta adapter with pstore implementation
moneta seems to be an overkill for our use case. we can do well with pstore instead for storing the key value pairs into file while staying compatible with faraday's http cache interface. this change also adds the pstore gem to gemspec and pins it to ~>0.1.1 because that's the version that was shipped with ruby 3.1.*.

Signed-off-by: Rishi Kumar Chawda <rishichawda@users.noreply.github.com>