add faraday middleware gem
chef-licensing was failing with a `'lookup_middleware': :json is not registered on Faraday::Request` error that was coming from the get connection method in rest client. faraday_middleware gems seems to be the one responsible to provide this json middleware https://github.com/lostisland/faraday_middleware/wiki/Changes-0.8. pinning faraday_middleware to ~>1.0.

Signed-off-by: Rishi Kumar Chawda <rishichawda@users.noreply.github.com>